Fully migrate layers slice, keeping legacy action type/shape
Consolidates the layers reducer and action creator into one
createSlice-based file (layersSlice.js), removing the separate
actions/layers.js. UPDATE_LAYERS stays in extraReducers with its
original mirador/UPDATE_LAYERS type and flat action shape rather than
becoming a slice-generated action — that type is exposed publicly via
ActionTypes, and third-party plugin sagas may pattern-match or
destructure fields directly off it (see the plugin ecosystem's use of
ActionTypes.* + takeEvery). Renaming/reshaping it would silently break
those plugins with no error, just a saga that stops firing.
